+++ CHANGELOG.md
@@ -4,12 +4,14 @@ All notable changes to Pipelines will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [0.0.14] - 2018-04-18
+### Added
+- Tag script to make releases
+### Changed
+- More useful default BITBUCKET_REPO_SLUG value
 ### Fixed
 - Coverage checker script precision
 - Duplicate output of non-zero exit code information
-### Added
-- Tag script to make releases
 
 ## [0.0.13] - 2018-03-20
 ### Fixed